progress made

+++ Controllers/EstateController.cs
@@ -1,9 +1,12 @@
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Http;
+
 using Newtonsoft.Json;
 
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.IO;
 
 using backend.DataAccess.Entities;
 using backend.Models;
@@ -31,10 +34,19 @@ namespace backend.Controllers
 
                 Advertisement ad = repo.GetAdvertisement(estate.Id);
                 Address address = repo.GetAddress(estate.Id);
-                GenericImage image = repo.GetEstateThumbnail(estate.Id);
+                GenericImage thumbnail = repo.GetEstateThumbnail(estate.Id);
+
+                if (ad == null || address == null)
+                    continue;
+
+                if (thumbnail == null) 
+                {
+                    thumbnail = new GenericImage();
+                    thumbnail.Id = "no-image";
+                    thumbnail.Extension = ".jpg";
+                }
 
                 estateOverview.EstateId = estate.Id;
-                estateOverview.Image = image.Id;
                 estateOverview.Title = ad.Title;
                 estateOverview.Price = estate.Price;
                 estateOverview.OfferType = ad.OfferType;
@@ -42,6 +54,8 @@ namespace backend.Controllers
                 estateOverview.PostCode = address.PostCode;
                 estateOverview.Street = address.Street;
                 estateOverview.Country = address.Country;
+                estateOverview.Image = thumbnail.Id + thumbnail.Extension;
+                estateOverview.TimePosted = ad.TimePosted;
 
                 response.Add(estateOverview);
             }
@@ -57,30 +71,22 @@ namespace backend.Controllers
             var response = new EstateDetailsViewModel();
 
             Estate estate = repo.GetEstate(estateId);
-            response.SquareFeet = estate.SquareFeet;
-            response.Category = estate.Category;
-            response.BuiltAt = estate.BuiltAt;
-            response.RefurbishedAt = estate.RefurbishedAt.ToString();
-            response.Grade = estate.Grade;
-            response.Room = estate.Room;
-            response.Kitchen = estate.Kitchen;
-            response.Bathroom = estate.Bathroom;
-            response.FloorCount = estate.FloorCount;
-            response.Garage = estate.Garage;
-            response.Elevator = estate.Elevator;
-            response.Terrace = estate.Terace;
-            response.PropertySquareFeet = estate.PropertySquareFeet;
-            response.GarageSquareFeet = estate.GarageSquareFeet ?? 0;
-            response.GardenSquareFeet = estate.GardenSquareFeet ?? 0;
-            response.TerraceSquareFeet = estate.TerraceSquareFeet ?? 0;
-            response.Basement = response.Basement;
-            response.Comfort = response.Comfort;
-            response.Advertiser = response.Advertiser;
-            response.AdvertiserId = response.AdvertiserId;
-            response.Images = repo.GetGenericImages()
-                                        .Where(image => image.EstateId == estate.Id)
-                                        .Select(image => image.Id)
-                                        .ToList();
+            Electricity electricity = repo.GetElectricity(estateId);
+            HeatingSystem heating = repo.GetHeatingSystem(estateId);
+            Address address = repo.GetAddress(estateId);
+            WaterSystem water = repo.GetWaterSystem(estateId);
+            PublicService services = repo.GetPublicService(estateId);
+            Advertisement advertisement = repo.GetAdvertisement(estateId);
+            List<GenericImage> images = repo.GetEstateImages(estateId);
+
+            response.Estate = estate;
+            response.Electricity = electricity;
+            response.Heating = heating;
+            response.Address = address;
+            response.Water = water;
+            response.Services = services;
+            response.Advertisement = advertisement;
+            response.Images = images;
 
             return JsonConvert.SerializeObject(response);
         }
@@ -136,6 +142,17 @@ namespace backend.Controllers
             return JsonConvert.SerializeObject(response);
         }
 
+        [HttpGet]
+        // Véletlenszerű képek.
+        //
+        public string RandomShowCase()
+        {
+            List<string> response = null;
+            response = repo.GetRandomImages(3);
+
+            return JsonConvert.SerializeObject(response);
+        }
+
         #endregion
         #region WebAPI POST
 
@@ -146,12 +163,19 @@ namespace backend.Controllers
         {
             var response = new ItemPostedResultModel();
 
+            if (!Startup.SignInManager.IsSignedIn(User)) {
+                response.Message = "Bejelentkezés szükséges.";
+                goto cancelUpload;
+            }
+
             var uploadable = new UploadWrapperModel();
             string data = null;
-            var  body = this.HttpContext.Request.Headers.ToArray();
+
             int length = (int) this.HttpContext.Request.ContentLength;
             byte[] buffer = new byte[length];
+            
             var bufferReaderTask = this.HttpContext.Request.Body.ReadAsync(buffer, 0, length);
+            bufferReaderTask.Wait();
 
             if (bufferReaderTask.IsCompletedSuccessfully)
             {
@@ -188,55 +212,52 @@ namespace backend.Controllers
             if (!uploadTask.IsCompletedSuccessfully)
                 response = new ItemPostedResultModel();
             
+            cancelUpload:
             return JsonConvert.SerializeObject(response);
         }
 
         [HttpPost]
         // Hirdetés módosítás.
         //
-        public string Modify(int estateId)
+        public string Modify()
         {
             var response = new ItemPostedResultModel();
 
-            var modifiable = new UploadWrapperModel();
-            string data = null;
-            var  body = this.HttpContext.Request.Headers.ToArray();
+            if (!Startup.SignInManager.IsSignedIn(User)) {
+                response.Message = "Bejelentkezés szükséges.";
+                goto cancelModify;
+            }
+
+            var model = new UploadWrapperModel();
+            
+            string data = null;            
             int length = (int) this.HttpContext.Request.ContentLength;
             byte[] buffer = new byte[length];
             var bufferReaderTask = this.HttpContext.Request.Body.ReadAsync(buffer, 0, length);
+            bufferReaderTask.Wait();
 
             if (bufferReaderTask.IsCompletedSuccessfully)
             {
                 data = System.Text.Encoding.Default.GetString(buffer);
-                modifiable = (UploadWrapperModel) JsonConvert.DeserializeObject(data, typeof(UploadWrapperModel), new JsonSerializerSettings());
+                model = (UploadWrapperModel) JsonConvert.DeserializeObject(data, typeof(UploadWrapperModel));
+
+                repo.UpdateEstate(model.Estate);
+                repo.UpdateAddress(model.Address);
+                repo.UpdateElectricity(model.Electricity);
+                repo.UpdateHeatingSystem(model.Heating);
+                repo.UpdatePublicService(model.PublicService);
+                repo.UpdateWaterSystem(model.Water);
+                repo.UpdateAdvertisement(model.Advertisement);
+
+                response.Message = "A hirdetés sikeresen módosítva lett.";
+                response.Success = true;
+                response.ItemId = model.Estate.Id.ToString();
+            }
+            else {
+                response.Message = "Hiba történt a hirdetés módosítása közben.";
             }
 
-            System.Func<object, bool> modifyEstateAdvertisementAsync = delegate(object toModify) 
-            {
-                var wrapper = (UploadWrapperModel) toModify;
-
-                repo.UpdateEstate(wrapper.Estate);
-                repo.UpdateAddress(wrapper.Address);
-                repo.UpdateElectricity(wrapper.Electricity);
-                repo.UpdateHeatingSystem(wrapper.Heating);
-                repo.UpdatePublicService(wrapper.PublicService);
-                repo.UpdateWaterSystem(wrapper.Water);
-                repo.UpdateAdvertisement(wrapper.Advertisement);
-
-
-                response.ItemId = wrapper.Estate.Id.ToString();
-                response.Success = true;
-
-                return true;
-            };
-
-            var modifyTask = new Task<bool>(modifyEstateAdvertisementAsync, modifiable);
-            modifyTask.Start();
-            modifyTask.Wait();
-
-            if (!modifyTask.IsCompletedSuccessfully)
-                response = new ItemPostedResultModel();
-
+            cancelModify:
             return JsonConvert.SerializeObject(response);
         }
 
@@ -247,8 +268,73 @@ namespace backend.Controllers
         {
             var response = new ItemPostedResultModel();
 
+            if (!Startup.SignInManager.IsSignedIn(User)) {
+                response.Message = "Bejelentkezés szükséges.";
+                goto cancelDelete;
+            }
+
             response.Success = repo.DeleteEstate(estateId);
             response.ItemId = estateId.ToString(); 
+
+            cancelDelete:
+            return JsonConvert.SerializeObject(response);
+        }
+
+        [HttpPost]
+        // Képfeltöltés
+        //
+        public string ImageUpload()
+        {
+            var response = new ItemPostedResultModel();
+
+            var images = HttpContext.Request.Form.Files;
+
+            long size = images.Sum(f => f.Length);
+            foreach (IFormFile image in images)
+            {
+                if (image.Length > 0)
+                {
+                    string cd = Directory.GetCurrentDirectory();
+                    string id = System.Guid.NewGuid().ToString();
+                    string extension = Path.GetExtension(image.FileName);
+                    string path = Path.Combine(cd, "Client", "static", id);
+                    path += extension;
+
+                    var imageData = new GenericImage();
+                    string json = Request.Form.ToDictionary(x => x.Key, x => x.Value.ToString())["imageData"];
+                    imageData = (GenericImage) JsonConvert.DeserializeObject(json, typeof(GenericImage));
+
+                    using (FileStream stream = System.IO.File.Create(path))
+                    {
+                        var copyImageToStreamTask = image.CopyToAsync(stream);
+                        copyImageToStreamTask.Wait();
+
+                        if (copyImageToStreamTask.IsCompletedSuccessfully)
+                        {
+                            imageData.Id = id;
+                            imageData.StorageSize = (int) image.Length;
+                            
+                            imageData.UserId = imageData.UserId;
+                            imageData.EstateId = imageData.EstateId;
+                            imageData.Category = imageData.Category;
+                            imageData.Title = imageData.Title;
+                            imageData.DescriptionDetail = imageData.DescriptionDetail;
+                            imageData.Extension = extension;
+
+                            bool storeSuccess = repo.AddImage(imageData);
+
+                            if (storeSuccess) {
+                                response.ItemId = imageData.Id;
+                                response.Success = true;
+                                response.Message = "A képfeltöltés sikeresen megtörtént.";
+                            }
+                        }                        
+                        else {
+                            response.Message = "Sikertelen képfeltöltés.";
+                        }
+                    }
+                }
+            }
 
             return JsonConvert.SerializeObject(response);
         }
